+++ app/page.tsx
@@ -1,4 +1,8 @@
 import Button from "@/components/shared/Button";
+import Link from "next/link";
+import Image from "next/image";
+import Discord from "@/assets/images/Discord.svg";
+import LinkedIn from "@/assets/images/LinkedIn.svg";
 
 const buttons = [
   {
@@ -21,31 +25,49 @@ const buttons = [
   },
   {
     id: 4,
-    title: "Join our Discord",
-    link: "https://www.codeandcoffee.chat",
-    color: "bg-purple-600",
-  },
-  {
-    id: 5,
     title: "Photos from past events",
     link: "/pastEvents",
     color: "bg-cyan-500",
   },
   {
-    id: 6,
+    id: 5,
     title: "About",
     link: "/about",
     color: "bg-blue-500",
   },
 ];
 
+const social_media = [
+  {
+    id : 1,
+    title: "Discord",
+    link: "https://www.codeandcoffee.chat",
+    image: Discord,
+  },
+  {
+    id : 2,
+    title: "LinkedIn",
+    link: "https://www.linkedin.com/company/boston-code-and-coffee",
+    image: LinkedIn,
+  },
+];
+
 export default function Home() {
   return (
     <main className="mx-auto max-w-7xl px-4 min-w-full">
-      <div className="my-2 h-full flex flex-col justify-evenly">
+      <div className="my-2 flex flex-col justify-evenly">
         {buttons.map(({ id, link, color, title }) => (
           <Button key={id} title={title} link={link} color={color} />
         ))}
+      </div>
+      <div className="flex my-2 justify-center">
+        {
+          social_media.map(({id, title, link, image})=>(            
+            <Link className="me-2" href={link} key={id}> 
+              <Image src={image} alt={title} width="40" height="40"/>
+            </Link>            
+          ))
+        }
       </div>
     </main>
   );